+++ backend/apid/graphql/asset.go
@@ -1,6 +1,7 @@
 package graphql
 
 import (
+	v2 "github.com/sensu/sensu-go/api/core/v2"
 	"github.com/sensu/sensu-go/backend/apid/graphql/globalid"
 	"github.com/sensu/sensu-go/backend/apid/graphql/schema"
 	"github.com/sensu/sensu-go/graphql"
@@ -18,6 +19,11 @@ func (*assetImpl) ID(p graphql.ResolveParams) (string, error) {
 
 // IsTypeOf is used to determine if a given value is associated with the type
 func (*assetImpl) IsTypeOf(s interface{}, p graphql.IsTypeOfParams) bool {
-	_, ok := s.(*types.Asset)
+	_, ok := s.(*v2.Asset)
 	return ok
+}
+
+// ToJSON implements response to request for 'toJSON' field.
+func (*assetImpl) ToJSON(p graphql.ResolveParams) (interface{}, error) {
+	return types.WrapResource(p.Source.(v2.Resource)), nil
 }