+++ backend/apid/graphql/handler.go
@@ -1,6 +1,7 @@
 package graphql
 
 import (
+	v2 "github.com/sensu/sensu-go/api/core/v2"
 	"github.com/sensu/sensu-go/backend/apid/graphql/globalid"
 	"github.com/sensu/sensu-go/backend/apid/graphql/schema"
 	"github.com/sensu/sensu-go/graphql"
@@ -50,6 +51,11 @@ func (r *handlerImpl) Handlers(p graphql.ResolveParams) (interface{}, error) {
 func (*handlerImpl) IsTypeOf(s interface{}, p graphql.IsTypeOfParams) bool {
 	_, ok := s.(*types.Handler)
 	return ok
+}
+
+// ToJSON implements response to request for 'toJSON' field.
+func (*handlerImpl) ToJSON(p graphql.ResolveParams) (interface{}, error) {
+	return types.WrapResource(p.Source.(v2.Resource)), nil
 }
 
 //